refactor javascript and use modules and Add luxon libray for time
