+++ index.html
@@ -23,6 +23,12 @@
     </header>
     <!-- Main -->
     <main>
+      <!-- Date Time -->
+      <section>
+        <p class="date-container"></p>
+      </section>
+
+      <!-- Pages  -->
       <section class="page" id="home">
         <h2>All Awesome Books</h2>
         <div class="books-container"></div>
@@ -58,6 +64,7 @@
     <footer>
       <div class="footer">Copyright ...</div>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
+    <script src="https://moment.github.io/luxon/global/luxon.js"></script>
   </body>
 </html>
